Move Windows compatibility layer to libnative
